refactor(site): add source import alias
