+++ src/site/pages/blog/article/page.ts
@@ -1,12 +1,11 @@
-import type { ArticleMetadata, BlogPost } from '../../../../content/index.ts';
-import type { SiteAssets } from '../../../assets.ts';
-import type { GeneratedFile } from '../../../generated-file.ts';
+import type { ArticleMetadata, BlogPost } from '@/content/index.ts';
+import { formatDate } from '@/lib/util.ts';
+import { islandModuleIds, type SiteAssets } from '@/site/assets.ts';
+import { SITE_ORIGIN } from '@/site/consts.ts';
+import type { GeneratedFile } from '@/site/generated-file.ts';
+import { page, renderComponent } from '@/site/html.ts';
+import { SITE_OWNER } from '@/site/site-owner.ts';
 import * as ufo from 'ufo';
-import { formatDate } from '../../../../lib/util.ts';
-import { islandModuleIds } from '../../../assets.ts';
-import { SITE_ORIGIN } from '../../../consts.ts';
-import { page, renderComponent } from '../../../html.ts';
-import { SITE_OWNER } from '../../../site-owner.ts';
 import path from 'node:path';
 import Article from './index.tsx';
