+++ src/site/pages/home/index.tsx
@@ -1,5 +1,5 @@
-import { SITE_ORIGIN } from '../../consts.ts';
-import ProfileImage from '../../components/ProfileImage/index.tsx';
+import ProfileImage from '@/site/components/ProfileImage/index.tsx';
+import { SITE_ORIGIN } from '@/site/consts.ts';
 import styles from './Home.module.css';
 
 const socials = [
