+++ src/site/pages/works/media/page.ts
@@ -1,7 +1,7 @@
-import type { SiteAssets } from '../../../assets.ts';
-import type { PostListItem } from '../../../content.ts';
-import type { GeneratedFile } from '../../../generated-file.ts';
-import { page, renderComponent } from '../../../html.ts';
+import type { SiteAssets } from '@/site/assets.ts';
+import type { PostListItem } from '@/site/content.ts';
+import type { GeneratedFile } from '@/site/generated-file.ts';
+import { page, renderComponent } from '@/site/html.ts';
 import Media from './index.tsx';
 
 /**
