+++ src/site/pages/works/publications/page.ts
@@ -1,6 +1,6 @@
-import type { SiteAssets } from '../../../assets.ts';
-import type { GeneratedFile } from '../../../generated-file.ts';
-import { page, renderComponent } from '../../../html.ts';
+import type { SiteAssets } from '@/site/assets.ts';
+import type { GeneratedFile } from '@/site/generated-file.ts';
+import { page, renderComponent } from '@/site/html.ts';
 import Publications from './index.tsx';
 
 type Publication = { title: string; link: string; authors: string; publisher: string };
