+++ src/site/pages/works/showcase/page.ts
@@ -1,7 +1,7 @@
-import type { ShowcaseProject } from '../../../../content/index.ts';
-import type { SiteAssets } from '../../../assets.ts';
-import type { GeneratedFile } from '../../../generated-file.ts';
-import { page, renderComponent } from '../../../html.ts';
+import type { ShowcaseProject } from '@/content/index.ts';
+import type { SiteAssets } from '@/site/assets.ts';
+import type { GeneratedFile } from '@/site/generated-file.ts';
+import { page, renderComponent } from '@/site/html.ts';
 import Showcase from './index.tsx';
 
 /**
