+++ src/site/pages/works/talks/page.ts
@@ -1,7 +1,7 @@
-import type { SiteAssets } from '../../../assets.ts';
-import type { GeneratedFile } from '../../../generated-file.ts';
-import type { Talk } from '../../../sections.ts';
-import { page, renderComponent } from '../../../html.ts';
+import type { SiteAssets } from '@/site/assets.ts';
+import type { GeneratedFile } from '@/site/generated-file.ts';
+import { page, renderComponent } from '@/site/html.ts';
+import type { Talk } from '@/site/sections.ts';
 import Talks from './index.tsx';
 
 /**
